requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
blake2b-py==0.1.4
cached-property==1.5.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.11.0
Django==3.2.6
djangorestframework==3.12.4
drf-yasg==1.20.0
eth-abi==2.1.1
eth-account==0.5.5
eth-bloom==1.0.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-tester==0.5.0b4
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
ipfshttpclient==0.7.0
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.0.1
model-bakery==1.3.2
multiaddr==0.0.9
multidict==5.1.0
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==21.0
parsimonious==0.8.1
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
py-ecc==4.1.0
py-evm==0.4.0a4
py-geth==3.4.0
pycryptodome==3.10.1
pyethash==0.1.27
pyparsing==2.4.7
pyrsistent==0.18.0
pysha3==1.0.2
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
requests==2.26.0
rlp==2.0.1
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
semantic-version==2.8.5
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.1
toml==0.10.2
toolz==0.11.1
trie==2.0.0a5
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
varint==1.0.2
web3==5.23.0
websockets==9.1
yarl==1.6.3
